+++ lib/main.dart
@@ -1,101 +1,14 @@
-import 'package:dev_doctor/pages/courses.dart';
-import 'package:dev_doctor/pages/editor.dart';
-import 'package:dev_doctor/pages/home.dart';
-import 'package:dev_doctor/pages/settings.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_localizations/flutter_localizations.dart';
+import 'package:flutter_modular/flutter_modular.dart';
 import 'package:hive/hive.dart';
 import 'package:hive_flutter/hive_flutter.dart';
 
-import 'generated/l10n.dart';
+import 'app_module.dart';
 
 void main() async {
   await Hive.initFlutter();
   var settingsBox = await Hive.openBox('settings');
   if (!settingsBox.containsKey('servers'))
     settingsBox.put('servers', ['https://backend.dev-doctor.cf']);
-  runApp(MyApp());
-}
-
-class MyApp extends StatelessWidget {
-  // This widget is the root of your application.
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      localizationsDelegates: [
-        S.delegate,
-        // ... app-specific localization delegate[s] here
-        // TODO: uncomment the line below after codegen
-        // AppLocalizations.delegate,
-        GlobalMaterialLocalizations.delegate,
-        GlobalWidgetsLocalizations.delegate,
-        GlobalCupertinoLocalizations.delegate,
-      ],
-      title: 'Dev-Doctor',
-      supportedLocales: S.delegate.supportedLocales,
-      themeMode: ThemeMode.light,
-      debugShowCheckedModeBanner: false,
-      theme: ThemeData(
-        fontFamily: "Roboto",
-        // This is the theme of your application.
-        //
-        // Try running your application with "flutter run". You'll see the
-        // application has a blue toolbar. Then, without quitting the app, try
-        // changing the primarySwatch below to Colors.green and then invoke
-        // "hot reload" (press "r" in the console where you ran "flutter run",
-        // or simply save your changes to "hot reload" in a Flutter IDE).
-        // Notice that the counter didn't reset back to zero; the application
-        // is not restarted.
-        /* appBarTheme: AppBarTheme(
-            color: Colors.white,
-            iconTheme: IconThemeData(color: Colors.black87),
-            textTheme: Theme.of(context).textTheme.merge(Typography.material2018().black).apply()), */
-        primarySwatch: Colors.red,
-      ),
-      home: MyHomePage(),
-    );
-  }
-}
-
-class MyHomePage extends StatefulWidget {
-  @override
-  _MyHomePageState createState() => _MyHomePageState();
-}
-
-class _MyHomePageState extends State<MyHomePage> {
-  int _selectedIndex = 0;
-  List<Widget> _widgetOptions = <Widget>[HomePage(), CoursesPage(), EditorPage(), SettingsPage()];
-
-  void _onItemTapped(int index) {
-    setState(() {
-      _selectedIndex = index;
-    });
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      body: Center(
-        child: _widgetOptions.elementAt(_selectedIndex),
-      ),
-      bottomNavigationBar: BottomNavigationBar(
-          type: BottomNavigationBarType.fixed,
-          items: const <BottomNavigationBarItem>[
-            BottomNavigationBarItem(
-                activeIcon: Icon(Icons.home), icon: Icon(Icons.home_outlined), label: 'Home'),
-            BottomNavigationBarItem(
-                activeIcon: Icon(Icons.school),
-                icon: Icon(Icons.school_outlined),
-                label: 'Courses'),
-            BottomNavigationBarItem(
-                activeIcon: Icon(Icons.create), icon: Icon(Icons.create_outlined), label: 'Editor'),
-            BottomNavigationBarItem(
-                activeIcon: Icon(Icons.settings),
-                icon: Icon(Icons.settings_outlined),
-                label: 'Settings')
-          ],
-          currentIndex: _selectedIndex,
-          onTap: _onItemTapped),
-    );
-  }
+  runApp(ModularApp(module: AppModule()));
 }
