+++ pubspec.yaml
@@ -30,7 +30,7 @@ dependencies:
   http: ^0.12.2
   hive: ^1.4.4+1
   hive_flutter: ^0.3.1
-  flutter_modular: ^2.0.1
+  flutter_modular: ^3.0.0-nullsafety.36
 
 
   # The following adds the Cupertino Icons font to your application.
